Initial commit of updated Digit Adaptor
This is a crazy container adaptor that treats integer types as
containers of digits in a particular radix.

This particular version of digit_adaptor has been upgraded somewhat
from versions that have previously appeared on Quora.  This version:

* eliminates the 'details' namespace,
* moves helper classes into digit_adaptor as private nested classes,
* eliminates use of depregated std::iterator<>
* renames digit_iterator and digit_proxy to iterator_ and reference_
* adds a pointer_ class

This isn't fully tested or by any means a fully complete container,
but it's much further along than the initial outing.  :-D