Clean up mutable/const reference and pointers
This splits out mutable_reference_ and const_reference_ into
separate types along with mutable_pointer_ and const_pointer_.

In the process, this also cleans up how we resolve swap so that it
actually works.  We need the specialization, as actual instances of
mutable_reference may themselves be const, in line with how real
references behave.